@@ -1,5 +1,6 @@
-var merge = require('webpack-merge')
-var devEnv = require('./dev.env')
+'use strict'
+const merge = require('webpack-merge')
+const devEnv = require('./dev.env')
 
 module.exports = merge(devEnv, {
   NODE_ENV: '"testing"'
